M4 review fixes: /expand run guard (#9) + bounded clustering memory (#11)
- #9: claim a run with an atomic guarded UPDATE (set status='running' only if
  not already running) via store.try_mark_running; /expand returns 409 if a run
  is in progress. Prevents a double-submit or retry-after-timeout (the prior run
  may still be executing server-side) from creating duplicate rows + double API
  spend. Race-safe, unlike a check-then-set.
- #11: cluster the similarity matrix in float32 (halves the O(n^2) footprint)
  and cap clustering input per topic to the top-N most-relevant actives
  (clustering_max_nodes, default 2500); the long-tail remainder stays active in
  the DB but unclustered. Bounds peak memory so a large active pool can't OOM
  the run.

+++ backend/app/pipeline/clustering.py
@@ -85,7 +85,9 @@ def cluster_topic(
     if n == 1:
         return [Grouping(f"{topic_id}:g0", [keywords[0]], keywords[0], 1.0, 1)]
 
-    vn = _normalized(np.asarray(embeddings, dtype=np.float64))
+    # float32 halves the n x n similarity matrix's footprint (the dominant
+    # allocation here); precision is ample for a 0.55 cosine threshold.
+    vn = _normalized(np.asarray(embeddings, dtype=np.float32))
     sims = vn @ vn.T  # cosine similarity matrix (rows are unit vectors)
 
     graph = nx.Graph()
